Update patch set 1
Patch Set 1:

In the interest of configurability, is there a way to make it serve up the robots.txt only if it already exists on the filesystem where it's looking? That way the robots.txt could be defined in our puppet config (as we tried to do last week and is in place now) and enabled/disabled/changed without digging into the source code.

Patch-set: 1
Reviewer: Gerrit User 6609 <6609@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0