Update patch set 1
Patch Set 1:

Has anyone run the tests for this project?  After installing the requirements and running `nosetests` they are all failing for me with:

 AttributeError: '_APIClientResponse' object has no attribute '_status_code'

I'm not sure if it's broken or I'm doing something wrong.

Note: alfajor has to be installed from source; the latest version is not on pypi.

Patch-set: 1
Label: Verified=0