Fix the tests by removing dependency on alfajor
The alfajor HTTP test wrapper is not working properly.  All the tests
are failing with:

  AttributeError: '_APIClientResponse' object has no attribute '_status_code'

It doesn't look like alfajor is being maintained any more [1, 2], so
remove the dependency on it and run the tests directly through the
werkzeug client instead.

Fix failure in test_json_get_styles by adding a missing style in the
expected results.

[1] https://github.com/idealist/Alfajor
[2] https://github.com/idealistdev/alfajor

Change-Id: Ie185ff37c5a13af0d1b5b59793a2cfb94d069069