Add libc6-dev to bindep
The python-builder base image was updated to no longer install
recommends. This is inline with the other Infra images and keeps
image sizes smaller. gcc recommended libc6-dev - but it turns out
we need that for stdio.h for one of our depends. Add it to fix
our image builds.

Change-Id: I18a808748979b7167361d2e5ebce04bc8eaf1347