Use less ambiguous line iteration variable name
In diff.py the line iteration inline function relied on an iterable
named "l" (lower-case L) which is easily confused with the number 1
in some poorly-designed fonts. Because this is the only occurrence
for which newer pycodestyle is objecting, just go ahead and adjust
it rather than altering our style checks.

Change-Id: I29cc8e5680d3fed3645b26f8b03d2ca299d974b0