Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Dmitriy Rabotyagov wrote:
> 
> > Maybe we can tag current state with smth like 0.1? What do you
> > think?
> 
> Good point, anything like a release of this codebase probably needs broader discussion. We didn't originally write lodgeit and its author intentionally deleted it from PyPI and GitHub years ago. There are also other forks floating around, like https://github.com/shirkey/spacepaste so I would definitely be worried about any attempts at making it look like an official release under the original lodgeit name.

I'd agree.  I don't see this having a life beyond simple containerisation.  i probably wouldn't bother but we archive a lot of the irc logs and they have a lot of paste references in them, so if we can keep it running with minimal effort it seems worth the trade-off.

Patch-set: 1