Update uwsgi image to python 3.10
OpenDev has stopped building python3.8 images at this point. This means
we need to update our Dockerfile to use a newer python version. In this
case python3.10 is chosen as it is newer but still easy to test on
Ubuntu. This will also update the version of uwsgi we are using.

Change-Id: I3eb4c580d854ebe0437114701d1b3f9c721e530e