Captcha: Fix text layer location calculation
ImageFont.getbbox() returns a tuple of coordinates for the text's
bounding box, (left, top, right, bottom).  Calculate the left margin
available in the image to place the text layer by subtracting the right
coordinate from the image width and adding it to the text's own left
margin. Similarly, to find the available top margin, subtract the bottom
coordinate from the image height and add it to the text's own top
margin.

Change-Id: I738816114be4c0a746733dc8e7ce5789aed57833