Reapply "Move lodgeit image publication to quay.io"
This reverts commit 5f6c89b5008ab5429b9ab67b5f535c10fd3fba33.

Now that paste.o.o is running on Noble with podman as the runtime we
should be able to get speculative container image testing when hosting
the images on quay. Move the lodgeit container image back to quay to
test all these assumptions.

Note that we need to mirror the uwsgi-base image to quay to fully remove
the dependency on docker hub. This is done in the depends on. Let's also
wait until we're confident with the Paste02 backups before committing
fully to the new docker less setup.

Depends-On: https://review.opendev.org/c/opendev/system-config/+/939383
Change-Id: Ia3b0b816c4b662ae41dcb163cb2aa65c65a13d94