Update lodgeit to use podman to build its container image
This is largely being done to make lodgeit sync up with system-config's
approach to image building. The plan in system-config is that when we
move images to quay we will build those images with podman as podman is
registry mirror aware. This enables speculative image testing even with
image build deps.

Change-Id: Ia8df7e2bcceb922cbe937de40caf927fcd7a1941